saved dictionary as yaml file
